Switch to using `usize`; fix number parsing
Numbers shouldn't have any letters directly afterwords. This is for groups that
start with a number.

+++ core/src/group.rs
@@ -20,7 +20,7 @@ impl Group {
         &self.ranges
     }
 
-    pub fn count(&self) -> u32 {
+    pub fn count(&self) -> usize {
         self.ranges.iter().map(|r| r.count()).sum()
     }
 }
